Use std::sqrt

+++ apps/probability/models/statistic/pooled_two_means_statistic.cpp
@@ -7,11 +7,11 @@ float PooledTwoMeansStatistic::computeDegreesOfFreedom(float s1, float n1, float
 }
 
 float PooledTwoMeansStatistic::computeStandardError(float s1, float n1, float s2, float n2) {
-  return _sp(s1, n1, s2, n2) * sqrtf(1. / n1  + 1. / n2);
+  return _sp(s1, n1, s2, n2) * std::sqrt(1. / n1 + 1. / n2);
 }
 
 float PooledTwoMeansStatistic::_sp(float s1, float n1, float s2, float n2) {
-  return sqrtf(((n1 - 1) * s1 * s1 + (n2 - 1) * s2 * s2) / (n1 + n2 - 2));
+  return std::sqrt(((n1 - 1) * s1 * s1 + (n2 - 1) * s2 * s2) / (n1 + n2 - 2));
 }
 
-} // namespace Probability
+}  // namespace Probability
